requirements.txt
locust>=2.40.0
python-dotenv>=1.0.0
